Split up the tests more logically
