+++ Tests/SpotHeroEmailValidatorTests/SpotHeroEmailValidatorTests.swift
@@ -17,20 +17,15 @@
 import XCTest
 
 class SpotHeroEmailValidatorTests: XCTestCase {
-    func testLevenshteinDistanceCategory() {
-        let tests = [
-            LevenshteinDistanceTestModel(stringA: "kitten", stringB: "sitting", distance: 3),
-            LevenshteinDistanceTestModel(stringA: "testing", stringB: "lev", distance: 6),
-            LevenshteinDistanceTestModel(stringA: "book", stringB: "back", distance: 2),
-            LevenshteinDistanceTestModel(stringA: "spot", stringB: "hero", distance: 4),
-            LevenshteinDistanceTestModel(stringA: "parking", stringB: "rules", distance: 6),
-            LevenshteinDistanceTestModel(stringA: "lame", stringB: "same", distance: 1),
-            LevenshteinDistanceTestModel(stringA: "same", stringB: "same", distance: 0),
-        ]
-
-        for test in tests {
-            XCTAssertEqual(test.stringA.levenshteinDistance(from: test.stringB), test.distance)
-            XCTAssertEqual(test.stringB.levenshteinDistance(from: test.stringA), test.distance)
+    struct ValidatorTestModel {
+        var emailAddress: String
+        var errorCode: UInt32
+        var suggestion: String?
+        
+        init(emailAddress: String, errorCode: UInt32 = 0, suggestion: String? = nil) {
+            self.emailAddress = emailAddress
+            self.errorCode = errorCode
+            self.suggestion = suggestion
         }
     }
