--- Tests/SpotHeroEmailValidatorTests/ValidatorTestModel.swift
@@ -1,28 +1,0 @@
-//  Copyright © 2019 SpotHero, Inc. All rights reserved.
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-//
-
-import Foundation
-
-struct ValidatorTestModel {
-    var emailAddress: String
-    var errorCode: UInt32
-    var suggestion: String?
-    
-    init(emailAddress: String, errorCode: UInt32 = 0, suggestion: String? = nil) {
-        self.emailAddress = emailAddress
-        self.errorCode = errorCode
-        self.suggestion = suggestion
-    }
-}
